backend/run.py
from dotenv import load_dotenv
import uvicorn

from app.core.config import get_settings


if __name__ == '__main__':
    load_dotenv()
    settings = get_settings()
    uvicorn.run('app.main:app', host=settings.api_host, port=settings.api_port, reload=True)
